+++ changelog.md
@@ -14,6 +14,7 @@
 * Removed the `PowerTransformer.getRating` overload which took a rating value. You can still get a rating via its `TransformerCoolingType`.
 * Removed `PowerTransformer.forEachRating` which looped over the collection with an index that made no sense. Please loop over `PowerTransformer.sRatings`
   instead.
+* `Equipment` to `EquipmentContainer` links for LV feeders are no longer written to the database, they should never have been.
 
 ### New Features
 * A file named after the ID of an ingestion job is now created when running `MetricsDatabaseWriter.save()`. For this feature to take effect, a `modelPath` must